refactor: remove top level generic "utils" file
