+++ src/puyapy/compile.py
@@ -31,7 +31,6 @@ from puyapy.awst_build.main import transform_ast
 from puyapy.client_gen import parse_arc56
 from puyapy.options import PuyaPyOptions
 from puyapy.parse import TYPESHED_PATH, ParseResult, parse_and_typecheck
-from puyapy.utils import determine_out_dir
 
 # this should contain the lowest version number that this compiler does NOT support
 # i.e. the next minor version after what is defined in stubs/pyproject.toml:tool.poetry.version
@@ -254,3 +253,31 @@ def _output_awst_any(
     output_path = awst_out_dir / f"module{suffix}"
     logger.info(f"writing {make_path_relative_to_cwd(output_path)}")
     output_path.write_text(out_text, "utf-8")
+
+
+def determine_out_dir(contract_path: Path, options: PuyaPyOptions) -> Path:
+    if not options.out_dir:
+        out_dir = contract_path
+    else:
+        # find input path the contract is relative to
+        for src_path in options.paths:
+            src_path = src_path.resolve()
+            src_path = src_path if src_path.is_dir() else src_path.parent
+            try:
+                relative_path = contract_path.relative_to(src_path)
+            except ValueError:
+                continue
+            # construct a path that maintains a hierarchy to src_path
+            out_dir = options.out_dir / relative_path
+            if not options.out_dir.is_absolute():
+                out_dir = src_path / out_dir
+            break
+        else:
+            # if not relative to any input path
+            if options.out_dir.is_absolute():
+                out_dir = options.out_dir / contract_path
+            else:
+                out_dir = contract_path / options.out_dir
+
+    out_dir.mkdir(parents=True, exist_ok=True)
+    return out_dir
